Begin abstracting ERK to work for all de types
