+++ src/dde/methods/mod.rs
@@ -4,4 +4,4 @@ pub use bs23::BS23; // Bogacki-Shampine 2(3) method with adaptive step size and 
 mod dopri5;
 pub use dopri5::DOPRI5; // Dormand-Prince 5(4) method with adaptive step size
 
-mod h_init;
+pub mod h_init;
