+++ src/methods/mod.rs
@@ -1,0 +1,4 @@
+//! Numerical Methods for Differential Equations
+
+mod erk;
+pub use erk::ExplicitRungeKutta;
