+++ src/ode/methods/runge_kutta/explicit/mod.rs
@@ -30,6 +30,3 @@ pub mod dopri5;
 
 pub use dop853::DOP853;
 pub use dopri5::DOPRI5;
-
-mod explicit_runge_kutta;
-pub use explicit_runge_kutta::ExplicitRungeKutta;
